Move definition search tool into /tool/app
