+++ settings.gradle.kts
@@ -16,6 +16,7 @@ include("game")
 include("cache")
 include("engine")
 include("tools")
+include("tools:app")
 include("buffer")
 include("network")
 include("types")
