+++ tools/app/src/main/composeResources/drawable/void_icon.png
