+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/Utils.kt
