+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/global/GlobalResult.kt
