+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/global/GlobalSearchTab.kt
@@ -56,12 +56,12 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import world.gregs.void.tools.generated.resources.Res
-import world.gregs.void.tools.generated.resources.arrow_drop_down
-import world.gregs.void.tools.generated.resources.arrow_right
-import world.gregs.void.tools.generated.resources.close
-import world.gregs.void.tools.generated.resources.open_in_new
-import world.gregs.void.tools.generated.resources.search
+import world.gregs.void.app.generated.resources.Res
+import world.gregs.void.app.generated.resources.arrow_drop_down
+import world.gregs.void.app.generated.resources.arrow_right
+import world.gregs.void.app.generated.resources.close
+import world.gregs.void.app.generated.resources.open_in_new
+import world.gregs.void.app.generated.resources.search
 import world.gregs.voidps.cache.Definition
 import world.gregs.voidps.tools.search.AccentBlue
 import world.gregs.voidps.tools.search.AccentLight
