+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/detail/BlockDetailRow.kt
