+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/detail/DetailLayout.kt
