+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/detail/ParamsDetail.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.jetbrains.compose.resources.painterResource
-import world.gregs.void.tools.generated.resources.Res
-import world.gregs.void.tools.generated.resources.open_in_new
+import world.gregs.void.app.generated.resources.Res
+import world.gregs.void.app.generated.resources.open_in_new
 import world.gregs.voidps.cache.Definition
 import world.gregs.voidps.tools.search.LinkColor
 import world.gregs.voidps.tools.search.ParamKey
