+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/tab/DefinitionTabContent.kt
@@ -56,8 +56,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.painterResource
-import world.gregs.void.tools.generated.resources.Res
-import world.gregs.void.tools.generated.resources.close
+import world.gregs.void.app.generated.resources.Res
+import world.gregs.void.app.generated.resources.close
 import world.gregs.voidps.cache.Definition
 import world.gregs.voidps.tools.search.AccentBlue
 import world.gregs.voidps.tools.search.AccentLight
