+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/table/ColumnHeader.kt
@@ -33,8 +33,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.jetbrains.compose.resources.painterResource
-import world.gregs.void.tools.generated.resources.Res
-import world.gregs.void.tools.generated.resources.arrow_drop_down
+import world.gregs.void.app.generated.resources.Res
+import world.gregs.void.app.generated.resources.arrow_drop_down
 import world.gregs.voidps.tools.search.AccentBlue
 import world.gregs.voidps.tools.search.BgPanel
 import world.gregs.voidps.tools.search.BorderColor
