+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/table/ResultRow.kt
@@ -36,8 +36,8 @@ import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.jetbrains.compose.resources.painterResource
-import world.gregs.void.tools.generated.resources.Res
-import world.gregs.void.tools.generated.resources.content_copy
+import world.gregs.void.app.generated.resources.Res
+import world.gregs.void.app.generated.resources.content_copy
 import world.gregs.voidps.cache.Definition
 import world.gregs.voidps.tools.search.AccentBlue
 import world.gregs.voidps.tools.search.BgMultiSelected
