+++ tools/app/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/table/filter/Filter.kt
